fix: solve duration rounding issue and show session complete in StartSessionCard only for today

+++ WGG/Models/Workout/SessionSet.swift
@@ -14,8 +14,8 @@ class SessionSet {
     var setNumber: Int
     var reps: Int
     var weight: Double
-    var setDuration: Double?
-    var restDuration: Double?
+    var setDuration: Int?
+    var restDuration: Int?
     
     var isCompleted: Bool = false
